requirements.txt
pytesseract==0.3.0
Flask==1.0.2
PyYAML==5.1